+++ src/app/views/Zap.svelte
@@ -1,46 +1,32 @@
-<script context="module" lang="ts">
-  type Zap = {
-    relay: string
-    pubkey: string
-    amount: number
-    status: string
-    relays?: string[]
-    invoice?: string
-    isTip?: boolean
-  }
-</script>
-
 <script lang="ts">
-  import {sortBy, uniqBy, uniq, whereEq, filter, map, reject} from "ramda"
-  import {doPipe} from "hurdak"
-  import {onDestroy} from "svelte"
+  import {init, launchPaymentModal, onModalClosed} from "@getalby/bitcoin-connect"
+  import {sortBy, uniqBy, filter, map, reject} from "ramda"
+  import {doPipe, Fetch} from "hurdak"
   import {now} from "@welshman/lib"
-  import {Tags} from "@welshman/util"
+  import {createEvent} from "@welshman/util"
+  import {generatePrivateKey} from "src/util/nostr"
   import Anchor from "src/partials/Anchor.svelte"
   import FieldInline from "src/partials/FieldInline.svelte"
   import Toggle from "src/partials/Toggle.svelte"
   import Input from "src/partials/Input.svelte"
   import Textarea from "src/partials/Textarea.svelte"
-  import ZapInvoice from "src/app/shared/ZapInvoice.svelte"
   import {router} from "src/app/util/router"
-  import {env, subscribe, hints, getSetting, derivePerson, requestZap} from "src/engine"
+  import {env, load, hints, signer, getSetting, derivePerson} from "src/engine"
 
   export let splits
   export let eid = null
   export let anonymous = false
   export let callback = null
 
-  let sub
-  let closing = false
-  let stage = "message"
   let message = ""
-  let zaps: Zap[] = []
+  let loading = false
+  let zaps = []
   let totalWeight = 0
   let totalAmount = getSetting("default_zap")
 
   $: {
     totalWeight = 0
-    zaps = doPipe<Zap[]>(splits, [
+    zaps = doPipe(splits, [
       reject((s: string[]) => s.length < 4 || s[1].length !== 64 || !s[3].match(/\d+(\.\d+)?$/)),
       map((s: string[]) => [...s.slice(1, 3), parseFloat(s[3])]),
       filter((s: any[]) => s[2] && s[2] > 0),
@@ -56,8 +42,8 @@
         amount: Math.round(totalAmount * (parseFloat(weight) / totalWeight)),
         status: "pending",
       })),
-      sortBy((split: Zap) => -split.amount),
-      (zaps: Zap[]) => {
+      sortBy((split: any) => -split.amount),
+      (zaps: any[]) => {
         const percent = getSetting("platform_zap_split")
 
         // Add our platform split on top as a "tip"
@@ -71,114 +57,100 @@
           })
         }
 
-        return zaps
+        // Add our zapper and relay hints
+        return zaps.map((zap, i) => {
+          const content = i === 0 ? message : ""
+          const {zapper} = derivePerson(zap.pubkey).get()
+          const relays = hints
+            .merge([hints.PublishMessage(zap.pubkey), hints.fromRelays([zap.relay])])
+            .getUrls()
+
+          return {...zap, zapper, relays, content}
+        })
       },
+      filter((zap: any) => zap.zapper?.lnurl),
+      // Request our invoice
+      map(async (zap: any) => {
+        const {amount, zapper, relays, content, pubkey} = zap
+        const msats = amount * 1000
+        const tags = [
+          ["relays", ...relays],
+          ["amount", msats.toString()],
+          ["lnurl", zapper.lnurl],
+          ["p", pubkey],
+        ]
+
+        if (eid) {
+          tags.push(["e", eid])
+        }
+
+        if (anonymous) {
+          tags.push(["anon"])
+        }
+
+        const template = createEvent(9734, {content, tags})
+        const signedTemplate = anonymous
+          ? await signer.get().signWithKey(template, generatePrivateKey())
+          : await signer.get().signAsUser(template)
+        const zapString = encodeURI(JSON.stringify(signedTemplate))
+        const qs = `?amount=${msats}&nostr=${zapString}&lnurl=${zapper.lnurl}`
+        const res = await tryCath(() => Fetch.fetchJson(zapper.callback + qs))
+
+        return {...zap, invoice: res?.pr}
+      }),
     ])
   }
 
-  const next = async () => {
-    stage = "invoice"
+  const confirmZap = async () => {
+    // Show loading immediately
+    loading = true
 
-    zaps.forEach(async (zap, i) => {
-      const msg = i === 0 ? message : ""
-      const {zapper} = derivePerson(zap.pubkey).get()
+    const since = now() - 30
+    const preppedZaps = await Promise.all(zaps)
 
-      if (!zapper?.lnurl) {
-        zaps[i].status = "error:zapper"
-        return
+    // Close the router once we can show the next modal
+    router.pop()
+
+    for (const {invoice, relays, zapper, pubkey} of preppedZaps) {
+      if (!invoice) {
+        continue
       }
 
-      const relays = hints
-        .merge([hints.PublishMessage(zap.pubkey), hints.fromRelays([zap.relay])])
-        .getUrls()
+      launchPaymentModal({invoice})
 
-      zaps[i].relays = relays
-      zaps[i].invoice = await requestZap(msg, zap.amount, {
-        eid,
-        relays,
-        anonymous,
-        lnurl: zapper.lnurl,
-        pubkey: zap.pubkey,
-      })
-    })
-
-    const successfulZaps = zaps.filter(z => !z.status.startsWith("error"))
-    const allRelays = uniq(successfulZaps.flatMap(z => z.relays))
-    const authors = successfulZaps.map(z => derivePerson(z.pubkey).get().zapper.nostrPubkey)
-    const mentions = successfulZaps.map(z => z.pubkey)
-
-    sub = subscribe({
-      relays: allRelays,
-      filters: [{kinds: [9735], authors, "#p": mentions, since: now() - 30}],
-      onEvent: e => {
-        zaps = zaps.map(z => {
-          if (z.invoice === Tags.fromEvent(e).get("bolt11")?.value()) {
-            z.status = "success"
-          }
-
-          return z
+      await new Promise<void>(resolve => {
+        const unsub = onModalClosed(() => {
+          resolve()
+          unsub()
         })
+      })
 
-        callback?.(e)
-      },
-    })
-  }
-
-  const prev = () => {
-    zaps = zaps.map(zap => ({...zap, invoice: null, status: "pending"}))
-    stage = "message"
-  }
-
-  const done = () => {
-    router.pop()
-  }
-
-  $: {
-    if (!closing && zaps.length > 0 && zaps.every(whereEq({status: "success"}))) {
-      closing = true
-      done()
+      load({
+        relays,
+        onEvent: callback,
+        filters: [{kinds: [9735], authors: [zapper.nostrPubkey], "#p": [pubkey], since}],
+      })
     }
   }
 
-  onDestroy(() => {
-    sub?.close()
-  })
+  // Initialize bitcoin connect
+  init({appName: import.meta.env.VITE_APP_NAME})
 </script>
 
 {#if zaps.length > 0}
-  {#if stage === "message"}
-    <h1 class="staatliches text-2xl">Send a zap</h1>
-    <Textarea bind:value={message} placeholder="Send a message with your zap (optional)" />
-    <Input bind:value={totalAmount}>
-      <i slot="before" class="fa fa-bolt" />
-      <span slot="after" class="-mt-1">sats</span>
-    </Input>
-    <FieldInline>
-      <div slot="label" class="flex items-center justify-between">
-        <div class="flex items-center gap-2">
-          <i class="fa fa-eye-slash" /> Zap anonymously
-        </div>
-        <Toggle bind:value={anonymous} />
+  <h1 class="staatliches text-2xl">Send a zap</h1>
+  <Textarea bind:value={message} placeholder="Send a message with your zap (optional)" />
+  <Input bind:value={totalAmount}>
+    <i slot="before" class="fa fa-bolt" />
+    <span slot="after" class="-mt-1">sats</span>
+  </Input>
+  <FieldInline>
+    <div slot="label" class="flex items-center justify-between">
+      <div class="flex items-center gap-2">
+        <i class="fa fa-eye-slash" /> Zap anonymously
       </div>
-    </FieldInline>
-    <Anchor button accent on:click={next}>Zap!</Anchor>
-  {:else}
-    {@const partitionAt = Math.ceil(zaps.length * 0.3)}
-    <div class="grid gap-8" class:sm:grid-cols-5={zaps.length > 1}>
-      <div class="flex flex-col gap-8" class:sm:col-span-3={zaps.length > 1}>
-        {#each zaps.slice(0, partitionAt) as zap (zap.pubkey)}
-          <ZapInvoice {zap} />
-        {/each}
-      </div>
-      <div class="flex flex-col gap-8" class:sm:col-span-2={zaps.length > 1}>
-        {#each zaps.slice(partitionAt) as zap (zap.pubkey)}
-          <ZapInvoice {zap} />
-        {/each}
-      </div>
+      <Toggle bind:value={anonymous} />
     </div>
-    <div class="flex gap-2">
-      <Anchor button on:click={prev}><i class="fa fa-arrow-left" /> Back</Anchor>
-      <Anchor button accent class="flex-grow" on:click={done}>Done zapping</Anchor>
-    </div>
-  {/if}
+  </FieldInline>
+  <Anchor button accent {loading} on:click={confirmZap}>Zap!</Anchor>
 {/if}